requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pandas
reportlab
websockets
psycopg2